Rename utils module as cli
