+++ src/cli/setup.rs
